build(`devcontainer/postinstall`): check MD5 hash before extract; echo if already satisfy

+++ .devcontainer/postinstall.sh
@@ -3,6 +3,8 @@
 DEVCONTAINER_DIR="/workspaces/yomuyume/.devcontainer"
 MOLD_VERSION="2.34.1"
 DAV1D_VERSION="1.4.3"
+DAV1D_MD5=c6fd9302a28d8c8e41e9a658a2be2031
+MOLD_MD5=08d7304ea9f5e232a5c46a45f230b5db
 
 echo "\n==============="
 echo "= pnpm stuffs ="
@@ -16,11 +18,13 @@ echo "\n==================="
 echo "= fix git symlink ="
 echo "===================\n"
 if [ ! -f /usr/local/bin/git ]; then
-    if [ ! -f /usr/bin/git ]; then
+    if [ -f /usr/bin/git ]; then
+        ln -s /usr/bin/git /usr/local/bin/git
+    else
         echo "git is not installed"
-        exit 1
     fi
-    ln -s /usr/bin/git /usr/local/bin/git
+else
+    echo "git is already exist in /usr/local/bin"
 fi
 
 echo "\n=================="
@@ -29,8 +33,14 @@ echo "==================\n"
 if [ ! -d "./dav1d-$DAV1D_VERSION" ]; then
     cd $DEVCONTAINER_DIR
     curl -L -o dav1d-$DAV1D_VERSION.tar.gz https://code.videolan.org/videolan/dav1d/-/archive/$DAV1D_VERSION/dav1d-$DAV1D_VERSION.tar.gz
-    tar -xvf dav1d-$DAV1D_VERSION.tar.gz
-    rm -f dav1d-$DAV1D_VERSION.tar.gz
+    if [ "$(md5sum dav1d-$DAV1D_VERSION.tar.gz | awk '{print $1}')" = "$DAV1D_MD5" ]; then
+        tar -xvf dav1d-$DAV1D_VERSION.tar.gz
+        rm -f dav1d-$DAV1D_VERSION.tar.gz
+    else
+        echo "dav1d-$DAV1D_VERSION.tar.gz has been modified"
+    fi
+else
+    echo "already downloaded dav1d-$DAV1D_VERSION"
 fi
 
 echo "\n==============="
@@ -41,6 +51,8 @@ if [ ! -d $DEVCONTAINER_DIR/dav1d-$DAV1D_VERSION/build ]; then
     mkdir build && cd build
     meson setup --default-library=static ..
     ninja
+else
+    echo "already built dav1d-$DAV1D_VERSION"
 fi
 
 echo "\n================="
@@ -54,8 +66,14 @@ echo "=============================\n"
 if [ ! -d $DEVCONTAINER_DIR/mold-$MOLD_VERSION-x86_64-linux ]; then
     cd $DEVCONTAINER_DIR
     curl -L -o mold-$MOLD_VERSION-x86_64-linux.tar.gz https://github.com/rui314/mold/releases/download/v$MOLD_VERSION/mold-$MOLD_VERSION-x86_64-linux.tar.gz
-    tar -xvf mold-$MOLD_VERSION-x86_64-linux.tar.gz
-    rm -f mold-$MOLD_VERSION-x86_64-linux.tar.gz
+    if [ "$(md5sum mold-$MOLD_VERSION-x86_64-linux.tar.gz | awk '{print $1}')" = "$MOLD_MD5" ]; then
+        tar -xvf mold-$MOLD_VERSION-x86_64-linux.tar.gz
+        rm -f mold-$MOLD_VERSION-x86_64-linux.tar.gz
+    else
+        echo "mold-$MOLD_VERSION-x86_64-linux.tar.gz has been modified"
+    fi
+else
+    echo "already downloaded mold-$MOLD_VERSION-x86_64-linux"
 fi
 
 rm -f /root/cargo/config.toml
